Delete README.md
<p align="center">
  

    ![Issues](https://img.shields.io/github/issues/ayush-raj8/Krishi-Lens)
![Pull Requests](https://img.shields.io/github/issues-pr/ayush-raj8/Krishi-Lens)
![Forks](https://img.shields.io/github/forks/ayush-raj8/Krishi-Lens)
![Stars](https://img.shields.io/github/stars/ayush-raj8/Krishi-Lens)
[![License](https://img.shields.io/github/license/ayush-raj8/Krishi-Lens)](https://github.com/ayush-raj8/Krishi-Lens)

  </a>
## 👀 Introduction
Krishi Lens will be an Android Application that will help farmers to detect disease in crop just by using smartphone with a camera. This project aims to leverage deep learning algorithm to bring smart farming in India in very economical ways. Farmer just need a smartphone with camera to know the type of disease ( if any) and it's cure just by clicking a photo. 

## 👩‍💻 Technology Stack

#### **Tools**

<img alt="Python" src="https://img.shields.io/badge/Python-3776AB?style=for-the-badge&logo=python&logoColor=white"/> 
<img alt="Jupyter Notebook" src="https://img.shields.io/badge/Jupyter-F37626.svg?&style=for-the-badge&logo=Jupyter&logoColor=white"/> 
<img alt="Kotlin" src="https://img.shields.io/badge/HTML-239120?style=for-the-badge&logo=html5&logoColor=white"/> 
<img alt="Deep Learning" src="https://img.shields.io/badge/CSS-239120?&style=for-the-badge&logo=css3&logoColor=white"/> 
<img alt="Android" src="https://img.shields.io/badge/Flutter-02569B?style=for-the-badge&logo=flutter&logoColor=white"/> 

## 🚀 Quick Start :

#### Step 1: Forking the repository :

To work on an open-source project, you will first need to make your copy of the repository. To do this, you should fork the repository and then clone it so that you have a local working copy.

Get your own Fork/Copy of repository by clicking the `Fork` button in right upper corner.<br><br>
#### Step 2: Clone the Forked Repository

After the repository is forked, you can now clone it so that you have a local working copy of the codebase.

To make your local copy of the repository follow the steps:
- Open the Command Prompt
- Type this command:

##### For Android App Develpment:
```bash
$ git clone -b flutter https://github.com/<your-github-username>/Krishi-Lens
```

## Drop an ⭐ on the Github Repository if you like it.<br><br>